Laboratório V/requirements.txt
paho-mqtt
tk
tk-tools
Events